requirements.txt
streamlit
gspread
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
google-genai
